Upgrade profile command to TypeScript and prisma instead of airtable
