+++ package.json
@@ -41,7 +41,9 @@
     "type": "module",
     "devDependencies": {
         "@babel/plugin-transform-modules-commonjs": "^7.15.4",
+        "@types/express": "^4.17.13",
         "@types/node": "^17.0.21",
+        "@types/twitter": "^1.7.1",
         "jest": "^27.2.5",
         "prisma": "^3.10.0",
         "ts-node": "^10.7.0",
