+++ src/app.ts
@@ -1,0 +1,8 @@
+//run the twitch bot
+import './bot.js';
+
+//run the tweeter
+import './tweeter.js';
+
+//start the express server
+import './server';
