--- src/utils/helpers.js
@@ -1,11 +1,0 @@
-const isValidUrl = (str) => {
-    try {
-        const url = new URL(str)
-        return true
-    } catch (err) {
-        console.error(`${str} is not a valid url`)
-        return false
-    }
-}
-
-export { isValidUrl }
