+++ src/utils/helpers.ts
@@ -1,0 +1,11 @@
+const isValidUrl = (str: string): boolean => {
+  try {
+    new URL(str)
+    return true
+  } catch (err) {
+    console.error(`${str} is not a valid url`)
+    return false
+  }
+}
+
+export { isValidUrl }
